Updated templates for project for dev and experimental
