+++ templates/necto/dev/project_templates/1_dual_core/CMakeLists.txt
