+++ templates/necto/dev/project_templates/3_designer/CMakeLists.txt
