+++ templates/necto/dev/project_templates/3_designer/main_screen.h
