+++ templates/necto/dev/project_templates/3_designer/resource.mres
