+++ templates/necto/dev/project_templates/5_library_sdk/lib.c
