+++ templates/necto/experimental/project_templates/4_application_with_library/template.json
