--- templates/necto/experimental/project_templates/4_lcd/template.json
@@ -1,14 +1,0 @@
-{
-  "name": "LCD output project",
-  "type": "lcd",
-  "language": "C",
-  "description": "Create a mikroSDK 2.0 2x16 LCD-ready application in C language.",
-  "project_file": "CMakeLists.txt",
-  "open_in_editor": ["main.c"],
-  "icon": ":/other/lcd",
-  "files": ["CMakeLists.txt", "main.c"],
-  "files_included": "**CMakeLists.txt:** Build configuration for mikroSDK LCD project.\n**main.c:** A main file with ready-to-use LCD initialization logic.",
-  "outputs": "**.hex:** Hexadecimal binary for MCU flashing.\n**.bin:** Raw binary for various deployment options.",
-  "requirements": "mikroSDK 2.0\nLCD 2x16 connected to your board",
-  "legacy": false
-}
