+++ templates/necto/experimental/project_templates/5_library_sdk/lib.c
