--- templates/necto/experimental/project_templates/5_plot/CMakeLists.txt
@@ -1,53 +1,0 @@
-cmake_minimum_required(VERSION 3.21)
-
-if (${TOOLCHAIN_LANGUAGE} MATCHES "MikroC")
-    project(${PROJECT_NAME} LANGUAGES MikroC)
-else()
-    project(${PROJECT_NAME} LANGUAGES C ASM)
-endif()
-
-include(mikroeUtils)
-
-if (${MIKROSDK_TYPE} STREQUAL "legacy")
-    legacy_cmake_path()
-endif()
-
-set(PROJECT_TYPE "${PROJECT_TYPE_VALUE}" CACHE STRING "" FORCE)
-
-add_executable(${PROJECT_NAME} main.c)
-############################ ${PROJECT_NAME} GENERATED CODE START ###########################
-############################ ${PROJECT_NAME} GENERATED CODE END ###########################
-
-find_package(MikroC.Core REQUIRED)
-find_package(MikroSDK.Plot REQUIRED)
-target_link_libraries(${PROJECT_NAME}
-    PUBLIC
-        MikroC.Core
-        MikroSDK.Plot
-)
-
-if (${MIKROSDK_TYPE} STREQUAL "legacy")
-    find_package(MikroC.System REQUIRED)
-    target_link_libraries(${PROJECT_NAME}
-        PUBLIC
-            MikroC.System
-            MikroSDK.Plot
-    )
-endif()
-
-if(COMPILER_REVISION)
-    if(${COMPILER_REVISION} VERSION_GREATER_EQUAL "1.0")
-        if(${TOOLCHAIN_LANGUAGE} STREQUAL "GNU")
-            if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/.meproject)
-                fetch_startup_name(${CMAKE_CURRENT_SOURCE_DIR}/.meproject startupName)
-                if(startupName)
-                    add_startup_all_targets(. ${CMAKE_CURRENT_SOURCE_DIR} ${startupName})
-                endif()
-                fetch_linker_name(${CMAKE_CURRENT_SOURCE_DIR}/.meproject linkerName)
-                if(linkerName)
-                    add_ld_all_targets(. ${CMAKE_CURRENT_SOURCE_DIR} ${linkerName})
-                endif()
-            endif()
-        endif()
-    endif()
-endif()
